refactor: remove __getattr__ from reminders/__init__.py, use direct imports
Reply to review: all SystemReminderService callers now import directly from
reminders.service instead of relying on module-level __getattr__.
