+++ src/relay_teams/agents/execution/agent_llm_session.py
@@ -54,7 +54,7 @@ from relay_teams.providers.model_fallback import (
     LlmFallbackMiddleware,
 )
 from relay_teams.providers.token_usage_repo import TokenUsageRepository
-from relay_teams.reminders import SystemReminderService
+from relay_teams.reminders.service import SystemReminderService
 from relay_teams.roles.memory_service import RoleMemoryService
 from relay_teams.roles.role_registry import RoleRegistry
 from relay_teams.sessions.runs.background_tasks import BackgroundTaskService
