+++ src/relay_teams/agents/orchestration/task_execution_service_factory.py
@@ -24,7 +24,7 @@ from relay_teams.agents.execution.message_repository import MessageRepository
 from relay_teams.sessions.runs.run_runtime_repo import RunRuntimeRepository
 from relay_teams.sessions.runs.run_intent_repo import RunIntentRepository
 from relay_teams.persistence.shared_state_repo import SharedStateRepository
-from relay_teams.reminders import SystemReminderService
+from relay_teams.reminders.service import SystemReminderService
 from relay_teams.agents.tasks.task_repository import TaskRepository
 from relay_teams.sessions.runs.todo_service import TodoService
 from relay_teams.skills.skill_registry import SkillRegistry
