+++ src/relay_teams/reminders/__init__.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from typing import Any
-
 from relay_teams.reminders.delivery import SystemReminderDeliveryMode
 from relay_teams.reminders.models import (
     CompletionAttemptObservation,
@@ -18,16 +16,6 @@ from relay_teams.reminders.state import ReminderStateRepository, ReminderRunStat
 from relay_teams.reminders.text import is_rendered_system_reminder_text
 from relay_teams.reminders.tool_effects import classify_tool_effect
 
-
-def __getattr__(name: str) -> Any:
-    if name == "SystemReminderService":
-        from relay_teams.reminders.service import SystemReminderService
-
-        globals()["SystemReminderService"] = SystemReminderService
-        return SystemReminderService
-    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
-
-
 __all__ = [
     "CompletionAttemptObservation",
     "ContextPressureObservation",
@@ -39,7 +27,6 @@ __all__ = [
     "ReminderStateRepository",
     "SystemReminderDeliveryMode",
     "SystemReminderPolicy",
-    "SystemReminderService",
     "ToolEffect",
     "ToolResultObservation",
     "classify_tool_effect",
